pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wdn"
version = "0.1.0"
description = "State reconstruction and anomaly detection in water distribution networks"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "WDN Thesis"}]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "torch>=2.1",
  "matplotlib>=3.7",
  "wntr>=1.2",
  "scikit-learn>=1.3",
  "tqdm>=4.66",
  "pytest>=7.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
